refactor(updates): extract recovery runtime
